removed/renamed generators
