+++ generators/custom-element/templates/index.js
