--- generators/element-next/templates/template.html
@@ -1,4 +1,0 @@
-<% if (postcss) { %><postcss src="./style.postcss"></postcss> <% } %>
-<section>
-
-</section>
