+++ generators/element/templates/methods/internal.js
