+++ generators/element/templates/methods/public.js
