+++ generators/element/templates/props.js
