+++ generators/element/templates/template.html
@@ -1,14 +1,4 @@
-<!-- Dependencies -->
+<% if (postcss) { %><postcss src="./style.postcss"></postcss> <% } %>
+<section>
 
-<!--
-`<<%= elementName %>>` is a starting template for new components .
-Basic use
-    <<%= elementName %>></<%= elementName %>>
-More comments here...
--->
-<dom-module id="<%= elementName %>">
-	<template>
-		<style></style>
-		
-	</template>
-</dom-module>
+</section>
