--- generators/element/templates/test.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-
-	<script src="./../../../webcomponentsjs/webcomponents-lite.js"></script>
-	<script src="./../../../../node_modules/web-component-tester/browser.js"></script>
-
-  <link rel="import" href="./../../../../node_modules/@polymer/polymer/polymer.html">
-	<link rel="import" href="./view.html">
-
-</head>
-<body>
-
-<test-fixture id="component">
-	<template>
-	  <<%= elementName %>></<%= elementName %>>
-	</template>
-</test-fixture>
-
-
-<script>
-suite('<<%= elementName %>>', function() {
-  var component;
-	setup(function() {
-		component = fixture('component');
-	});
-	suite('Test Properties', function() {
-		test('Default Properties', function() {
-			assert.isUndefined(component.prop);
-		});
-	});
-});
-</script>
-</body>
-</html>
