--- generators/readme/templates/readme.md
@@ -1,23 +1,0 @@
-# <%= projectTitle %>
-
-<!-- Description here -->
-
-### Requirements
-* Node v6.x
-
-
-### Installation & start
-* Clone this repo on your machine
-* On project root launch 
-```
-$ npm install
-```
-
-
-### Available tasks
-
-
-<% if (license) { %>
-### License
-Released under the [Apache 2.0 license](LICENSE).
-<% } %>
